Refactor: Move getPrivateImageUrlsCreatedAfter to VoicesRepository
Moved `getPrivateImageUrlsCreatedAfter` from `ResourcesRepository` to `VoicesRepository` and updated its usage in `NewsViewModel` and its corresponding test.
