+++ app/src/main/java/org/ole/planet/myplanet/repository/ResourcesRepositoryImpl.kt
@@ -294,15 +294,6 @@ class ResourcesRepositoryImpl @Inject constructor(
         return results.filter { it.needToUpdate() }
     }
 
-    override suspend fun getPrivateImageUrlsCreatedAfter(timestamp: Long): List<String> {
-        val imageList = queryList(RealmMyLibrary::class.java) {
-            equalTo("isPrivate", true)
-                .greaterThan("createdDate", timestamp)
-                .equalTo("mediaType", "image")
-        }
-        return imageList.mapNotNull { it.resourceRemoteAddress }
-    }
-
     override fun getRecentResources(userId: String): Flow<List<RealmMyLibrary>> {
         return queryListFlow(RealmMyLibrary::class.java) {
             equalTo("userId", userId)
